Moves logo.js to componenets from containers
