+++ client/src/components/logo/index.js
